Release/0.1.1 (#9)
* Set pump initial settings and pump popover

* Check reservoir

* fix initial setup

* open CGM source

* fix initial pump setup

* Hide debug options

* fix initial preferences

* use xcconfig

* Autoupdate build number

* Create README_RU.md

* Display version and build number

* English README

* Disclaimer

* disallow bolus 0

* possible fix wrong layout after start

* fix glucose Y range

* Optional api secret

* fix suggestion parsing

* Face id access

* fix clock

* build autoupdate

* Fix temp targets layout and pump view

* Fix pump updating logic

* fix omnipod reservooir

* Background timers

* legend

* Remove unused dependecies

* show temp target range

* FAQ_RU

* FAQ eng

* bump buils number

* clear signing

* fix reservoir bug

* Libre 2 glucose filter

* fix glucose filter logic

* skip_neutral_temps option fix

* Fix bild number update script

* Fix BloodGlucose parsing

* update charts on UIApplication.willEnterForegroundNotification

* Fix autotune deletion

* README updated

* Bump version

+++ FreeAPS/Sources/Models/BloodGlucose.swift
@@ -23,7 +23,7 @@ struct BloodGlucose: JSON, Identifiable, Hashable {
 
     var sgv: Int?
     let direction: Direction?
-    let date: UInt64
+    let date: Decimal
     let dateString: Date
     let filtered: Double?
     let noise: Int?
